Remove mock_options.[c|h]
Prefer test-file-specific options instead of one, global options definition. The
baremetal eventrouter does not support options that define multiple tasks, which
is required in FreeRTOS eventrouter tests; therefore there is no set of options
which can support all tests.
