--- extra/test/mock_options.cc
@@ -1,3 +1,0 @@
-#include "mock_options.h"
-
-bool MockOptions::m_is_in_isr = false;
